Split branding tests into module
